KriptosKratos/requirements.txt
ccxt>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
tensorflow>=2.7.0
scikit-learn>=0.24.0
pandas-ta>=0.3.0
python-dotenv>=0.19.0
colorama>=0.4.4
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
statsmodels>=0.13.0
